Create show.html.erb
Master did not have show.html.erb file. Not sure why. Copied the file and manually created it on this repo. Stemmed from continuous Heroku error messages. Heroku logs showed that there was a missing template. Then looked at Git realizing the template was not here. Suspect it's because of multiple upstream branches causing git to refuse to push. Was in essence pushing an incomplete file/app to Heroku.